requirements.txt
aiohttp==3.9.3
pandas==2.2.0
pyyaml==6.0.1
tenacity==8.2.3
pyarrow==15.0.0  # for parquet support
